refactor: move shadcn-ui components back to default folder
