+++ server/app/routes/_index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useRef, useState } from "react";
 import { TbSearch } from "react-icons/tb";
-import { Button } from "~/components/shadcn-ui/button";
+import { Button } from "~/components/ui/button";
 import type { ActionArgs } from "@remix-run/node";
 import { json, redirect } from "@remix-run/node";
 import { Form, useFetcher, useNavigate, useSubmit } from "@remix-run/react";
